requirements.txt
Flask
Pillow
pytesseract
pillow-heif
